requirements.txt


# AcaDoc AI POC - Requirements
# Linked to Technical Plan v1.2 Sections 3, 4, 8, 9
fastapi>=0.100.0
uvicorn>=0.20.0

# Core Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0

# PDF Processing
pypdf>=4.0.0
pymupdf>=1.23.0

# Vector Store & Embeddings
surrealdb>=1.0.0
langchain-surrealdb>=0.0.1
sentence-transformers>=3.0.0
huggingface-hub>=0.20.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# LLM Integration (OpenAI or Ollama)
openai>=1.0.0
ollama>=0.1.0
langchain-ollama>=0.1.0

# Validation & Types
pydantic>=2.0.0
pydantic-settings>=2.0.0

# UI
streamlit>=1.30.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.0
google-cloud-storage>=2.14.0